requirements.txt
gym==0.23.1
matplotlib
numpy
torch
seaborn
ruamel.yaml
pytest